rename Horn clause SAT/UNSAT results for use by all solvers
