+++ trunk/source/Library-UltimateTest/src/de/uni_freiburg/informatik/ultimate/test/decider/overallresult/TreeAutomizerOverallResultEvaluator.java
@@ -33,8 +33,8 @@ import java.util.Map.Entry;
 import java.util.Set;
 
 import de.uni_freiburg.informatik.ultimate.core.lib.results.TimeoutResult;
-import de.uni_freiburg.informatik.ultimate.core.lib.results.TreeAutomizerSatResult;
-import de.uni_freiburg.informatik.ultimate.core.lib.results.TreeAutomizerUnsatResult;
+import de.uni_freiburg.informatik.ultimate.core.lib.results.ChcSatResult;
+import de.uni_freiburg.informatik.ultimate.core.lib.results.ChcUnsatResult;
 import de.uni_freiburg.informatik.ultimate.core.model.results.IResult;
 import de.uni_freiburg.informatik.ultimate.core.model.services.IResultService;
 
@@ -45,18 +45,18 @@ public class TreeAutomizerOverallResultEvaluator implements IOverallResultEvalua
 
 	@Override
 	public void evaluateOverallResult(IResultService resultService) {
-		TreeAutomizerSatResult satResult = null;
-		TreeAutomizerUnsatResult unsatResult = null;
+		ChcSatResult satResult = null;
+		ChcUnsatResult unsatResult = null;
 		TimeoutResult timeoutResult = null;
 		for (Entry<String, List<IResult>> en : resultService.getResults().entrySet()) {
 			for (IResult res : en.getValue()) {
-				if (res instanceof TreeAutomizerSatResult) {
+				if (res instanceof ChcSatResult) {
 					assert satResult == null && unsatResult == null : "more than one (un)sat result";
-					satResult = (TreeAutomizerSatResult) res;
+					satResult = (ChcSatResult) res;
 					mMostSignificatResults.add(res);
-				} else if (res instanceof TreeAutomizerUnsatResult) {
+				} else if (res instanceof ChcUnsatResult) {
 					assert satResult == null && unsatResult == null : "more than one (un)sat result";
-					unsatResult = (TreeAutomizerUnsatResult) res;
+					unsatResult = (ChcUnsatResult) res;
 					mMostSignificatResults.add(res);
 				} else if (res instanceof TimeoutResult) {
 					timeoutResult = (TimeoutResult) res;
